Parameterize proxy/task related properties
Provide the ability to resolve the pinned servers property
from environment variables when deploying tasks/proxy services.

Related to: https://github.com/wso2/micro-integrator/issues/847

Docs: https://github.com/wso2/micro-integrator/issues/868
